requirements.txt
flask
faiss-cpu
sentence-transformers
requests
pandas
numpy
